New features: sign in with github and sign out.
Implement OAuth with uberauth library, following the documentation.

Create a new context called authentication to encapsulate OAuth and user
session operations related to the database.

Create the user entity inside the authentication context.

Create the 'set_user' plug to insert the current user into every
connection piped through the browser pipeline.

Create a logic to show a 'Sing Out' button when a user is logged in and
a 'Sign In with Github' button when no user is logged in.